Move FeedHqUrlReader into a separate file
