+++ include/feedhqurlreader.h
@@ -1,0 +1,30 @@
+#ifndef NEWSBOAT_FEEDHQURLREADER_H_
+#define NEWSBOAT_FEEDHQURLREADER_H_
+
+#include "urlreader.h"
+
+namespace newsboat {
+
+class ConfigContainer;
+class RemoteApi;
+
+class FeedHqUrlReader : public UrlReader {
+public:
+	FeedHqUrlReader(ConfigContainer* c,
+		const std::string& url_file,
+		RemoteApi* a);
+	~FeedHqUrlReader() override;
+	void write_config() override;
+	void reload() override;
+	std::string get_source() override;
+
+private:
+	ConfigContainer* cfg;
+	std::string file;
+	RemoteApi* api;
+};
+
+} // namespace newsboat
+
+#endif /* NEWSBOAT_FEEDHQURLREADER_H_ */
+
