+++ mk/mk.deps
@@ -93,15 +93,15 @@ src/controller.o: src/controller.cpp include/controller.h include/cache.h \
  include/colormanager.h include/configcontainer.h \
  include/configexception.h include/configparser.h include/dbexception.h \
  include/downloadthread.h include/exception.h include/feedhqapi.h \
- include/fileurlreader.h include/globals.h include/inoreaderapi.h \
- include/itemrenderer.h include/htmlrenderer.h include/textformatter.h \
- include/logger.h include/newsblurapi.h rss/feed.h rss/item.h \
- include/ocnewsapi.h include/oldreaderapi.h include/opmlurlreader.h \
- include/regexmanager.h include/remoteapi.h include/rssparser.h \
- include/stflpp.h include/strprintf.h include/ttrssapi.h \
- 3rd-party/json.hpp include/utils.h include/view.h include/controller.h \
- include/filebrowserformaction.h include/formaction.h include/history.h \
- include/keymap.h include/stflpp.h
+ include/feedhqurlreader.h include/fileurlreader.h include/globals.h \
+ include/inoreaderapi.h include/itemrenderer.h include/htmlrenderer.h \
+ include/textformatter.h include/logger.h include/newsblurapi.h \
+ rss/feed.h rss/item.h include/ocnewsapi.h include/oldreaderapi.h \
+ include/opmlurlreader.h include/regexmanager.h include/remoteapi.h \
+ include/rssparser.h include/stflpp.h include/strprintf.h \
+ include/ttrssapi.h 3rd-party/json.hpp include/utils.h include/view.h \
+ include/controller.h include/filebrowserformaction.h \
+ include/formaction.h include/history.h include/keymap.h include/stflpp.h
 src/dialogsformaction.o: src/dialogsformaction.cpp \
  include/dialogsformaction.h include/formaction.h include/history.h \
  include/keymap.h include/configparser.h include/configactionhandler.h \
@@ -134,14 +134,13 @@ src/feedhqapi.o: src/feedhqapi.cpp include/feedhqapi.h include/cache.h \
  include/configcontainer.h include/configparser.h \
  include/configactionhandler.h include/rss.h include/matcher.h \
  filter/FilterParser.h include/utils.h include/logger.h config.h \
- include/strprintf.h include/remoteapi.h include/urlreader.h \
- include/strprintf.h include/utils.h
-src/feedhqurlreader.o: src/feedhqurlreader.cpp include/feedhqapi.h \
- include/cache.h include/configcontainer.h include/configparser.h \
- include/configactionhandler.h include/rss.h include/matcher.h \
- filter/FilterParser.h include/utils.h include/logger.h config.h \
- include/strprintf.h include/remoteapi.h include/urlreader.h \
- include/fileurlreader.h include/logger.h
+ include/strprintf.h include/remoteapi.h include/strprintf.h \
+ include/utils.h
+src/feedhqurlreader.o: src/feedhqurlreader.cpp include/feedhqurlreader.h \
+ include/urlreader.h include/configcontainer.h include/configparser.h \
+ include/configactionhandler.h include/fileurlreader.h include/logger.h \
+ config.h include/strprintf.h include/remoteapi.h \
+ include/configcontainer.h include/utils.h include/logger.h
 src/feedlistformaction.o: src/feedlistformaction.cpp \
  include/feedlistformaction.h include/history.h include/listformaction.h \
  include/formaction.h include/keymap.h include/configparser.h \
